requirements.txt
google-adk
litellm
streamlit
google.genai
langchain_community
langchain_chroma
beautifulsoup4
langchain_openai